editor widget 개편
widget 대규모 개편
- data 이름을 출력하는 위젯들의 text widget을 분리
- data를 관리하는 widgetdata class 추가
기존 data를 다루는 위젯들의 부모 클래스로 변경 예정

- graphics setting gui 추가

graphics
- post process test 완료
- graphics setting 추가

+++ Editor/src/editor/GUI/ProjectView.cpp
@@ -430,7 +430,7 @@ namespace editor
 			1
 		};
 
-		graphics::ImageDesc _imageDesc{ _pixel.data() , _pitch, _pitch * _height};
+		graphics::ImageDesc _imageDesc{ _pixel.data() , (uint32)_pitch, (uint32)(_pitch * _height)};
 
 		auto _texture = rengine::Resources::GetInstance()->CreateResource<rengine::Texture>();
